requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-otp==0.9.0
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
jsmin==2.2.2
Pillow==7.1.2
psycopg2==2.8.5
pymongo==3.10.1
pyotp==2.3.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
yagmail==0.11.224
